requirements.txt
alabaster==0.7.12
asn1crypto==0.24.0
Babel==2.7.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
docutils==0.15.2
idna==2.8
imagesize==1.1.0
Jinja2==2.10.1
joblib==0.13.2
kaptan==0.5.12
MarkupSafe==1.1.1
mkl-fft==1.0.12
mkl-random==1.0.2
mkl-service==2.0.2
numpy==1.16.4
olefile==0.46
packaging==19.1
Pillow==6.1.0
pycparser==2.19
Pygments==2.4.2
pyOpenSSL==19.0.0
pyparsing==2.4.2
PySocks==1.7.0
pytz==2019.2
PyYAML==5.1.2
pyzmq==18.1.0
requests==2.22.0
scikit-learn==0.21.2
scipy==1.3.1
six==1.12.0
snowballstemmer==1.9.0
Sphinx==2.1.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
torch==1.2.0
torchfile==0.1.0
torchsummary==1.5.1
torchvision==0.4.0
tornado==6.0.3
tqdm==4.33.0
urllib3==1.25.3
visdom==0.1.8.8
websocket-client==0.56.0
win-inet-pton==1.1.0
wincertstore==0.2
yacs==0.1.6
